Updated msvc 200X projects, fixed some warnings and one typo
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@1411 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/CProjectile.cpp
@@ -272,7 +272,7 @@ int CProjectile::CheckCollision(float dt, CMap *map, CWorm* worms, float* enddt)
 	
 	// HINT: this prevents napalms and similar stuff from flying through walls (serious bug)
 	if (tProjInfo->Dampening <= 1) // This rule does not apply to "accelerating" projectiles
-		len = MIN(tProjInfo->ProjSpeed + tProjInfo->ProjSpeedVar * iRandom, len);
+		len = MIN(tProjInfo->ProjSpeed + (int)(tProjInfo->ProjSpeedVar * iRandom), len);
 
 	
 	if (len < 14000)  {
